feat: refactor to make integration tests more extensible
