+++ controllers/sleepinfo/sleepinfo_controller_test.go
@@ -3,7 +3,6 @@ package sleepinfo
 import (
 	"context"
 	"encoding/json"
-	"fmt"
 	"time"
 
 	kubegreenv1alpha1 "github.com/davidebianchi/kube-green/api/v1alpha1"
@@ -20,6 +19,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 )
 
+// TODO: add tests for cron jobs
 var _ = Describe("SleepInfo Controller", func() {
 	const (
 		sleepInfoName = "sleep-name"
@@ -45,7 +45,7 @@ var _ = Describe("SleepInfo Controller", func() {
 
 	namespace := "zero-deployments"
 	It("reconcile - zero deployments", func() {
-		createdSleepInfo := createSleepInfo(ctx, sleepInfoName, namespace, SetupOptions{})
+		createdSleepInfo := createSleepInfo(ctx, sleepInfoName, namespace, setupOptions{})
 
 		req := reconcile.Request{
 			NamespacedName: types.NamespacedName{
@@ -75,8 +75,7 @@ var _ = Describe("SleepInfo Controller", func() {
 		Expect(err).NotTo(HaveOccurred())
 
 		By("not exists deployments", func() {
-			deployments, err := listDeployments(ctx, namespace)
-			Expect(err).NotTo(HaveOccurred())
+			deployments := listDeployments(ctx, namespace)
 			Expect(len(deployments)).To(Equal(0))
 		})
 
@@ -116,8 +115,7 @@ var _ = Describe("SleepInfo Controller", func() {
 		Expect(err).NotTo(HaveOccurred())
 
 		By("no deployment in namespace", func() {
-			deployments, err := listDeployments(ctx, namespace)
-			Expect(err).NotTo(HaveOccurred())
+			deployments := listDeployments(ctx, namespace)
 			Expect(len(deployments)).To(Equal(0))
 		})
 
@@ -266,7 +264,7 @@ var _ = Describe("SleepInfo Controller", func() {
 
 	It("reconcile - with deployments", func() {
 		namespace := "multiple-deployments"
-		req, originalDeployments := setupNamespaceWithDeployments(ctx, sleepInfoName, namespace, sleepInfoReconciler, mockNow, SetupOptions{})
+		req, originalResources := setupNamespaceWithResources(ctx, sleepInfoName, namespace, sleepInfoReconciler, mockNow, setupOptions{})
 
 		assertContextInfo := AssertOperation{
 			testLogger:          testLogger,
@@ -274,7 +272,7 @@ var _ = Describe("SleepInfo Controller", func() {
 			req:                 req,
 			namespace:           namespace,
 			sleepInfoName:       sleepInfoName,
-			originalDeployments: originalDeployments,
+			originalDeployments: originalResources.deploymentList,
 		}
 
 		assertCorrectSleepOperation(assertContextInfo.withSchedule("2021-03-23T20:05:59.000Z").withRequeue(14*60 + 1))
@@ -284,7 +282,7 @@ var _ = Describe("SleepInfo Controller", func() {
 
 	It("reconcile - deploy between sleep and wake up", func() {
 		namespace := "deploy-between-sleep-and-wake-up"
-		req, originalDeployments := setupNamespaceWithDeployments(ctx, sleepInfoName, namespace, sleepInfoReconciler, mockNow, SetupOptions{})
+		req, originalResources := setupNamespaceWithResources(ctx, sleepInfoName, namespace, sleepInfoReconciler, mockNow, setupOptions{})
 
 		assertContextInfo := AssertOperation{
 			testLogger:          testLogger,
@@ -292,20 +290,18 @@ var _ = Describe("SleepInfo Controller", func() {
 			req:                 req,
 			namespace:           namespace,
 			sleepInfoName:       sleepInfoName,
-			originalDeployments: originalDeployments,
+			originalDeployments: originalResources.deploymentList,
 		}
 
 		assertCorrectSleepOperation(assertContextInfo.withSchedule("2021-03-23T20:05:59.000Z").withRequeue(14*60 + 1))
 
 		By("re deploy", func() {
-			_, err := upsertDeployments(ctx, namespace, true)
-			Expect(err).NotTo(HaveOccurred())
+			upsertDeployments(ctx, namespace, true)
 
 			By("check replicas")
-			deployments, err := listDeployments(ctx, namespace)
-			Expect(err).NotTo(HaveOccurred())
+			deployments := listDeployments(ctx, namespace)
 			for _, deployment := range deployments {
-				originalDeployment := findDeployByName(originalDeployments, deployment.Name)
+				originalDeployment := findDeployByName(originalResources.deploymentList, deployment.Name)
 				Expect(deployment.Spec.Replicas).To(Equal(originalDeployment.Spec.Replicas))
 			}
 		})
@@ -315,7 +311,7 @@ var _ = Describe("SleepInfo Controller", func() {
 
 	It("reconcile - change single deployment replicas between sleep and wake up", func() {
 		namespace := "change-replicas-between-sleep-and-wake-up"
-		req, originalDeployments := setupNamespaceWithDeployments(ctx, sleepInfoName, namespace, sleepInfoReconciler, mockNow, SetupOptions{})
+		req, originalResources := setupNamespaceWithResources(ctx, sleepInfoName, namespace, sleepInfoReconciler, mockNow, setupOptions{})
 
 		assertContextInfo := AssertOperation{
 			testLogger:          testLogger,
@@ -323,18 +319,17 @@ var _ = Describe("SleepInfo Controller", func() {
 			req:                 req,
 			namespace:           namespace,
 			sleepInfoName:       sleepInfoName,
-			originalDeployments: originalDeployments,
+			originalDeployments: originalResources.deploymentList,
 		}
 		assertCorrectSleepOperation(assertContextInfo.withSchedule("2021-03-23T20:05:59.000Z").withRequeue(14*60 + 1))
 
 		By("re deploy a single deployment", func() {
-			deployments, err := listDeployments(ctx, namespace)
-			Expect(err).NotTo(HaveOccurred())
+			deployments := listDeployments(ctx, namespace)
 
 			deploymentToUpdate := deployments[0].DeepCopy()
 			patch := client.MergeFrom(deploymentToUpdate)
 			*deploymentToUpdate.Spec.Replicas = 0
-			err = k8sClient.Patch(ctx, deploymentToUpdate, patch)
+			err := k8sClient.Patch(ctx, deploymentToUpdate, patch)
 			Expect(err).ToNot(HaveOccurred())
 
 			updatedDeployment := appsv1.Deployment{}
@@ -347,7 +342,7 @@ var _ = Describe("SleepInfo Controller", func() {
 
 	It("reconcile - twice consecutive sleep operation", func() {
 		namespace := "twice-sleep"
-		req, originalDeployments := setupNamespaceWithDeployments(ctx, sleepInfoName, namespace, sleepInfoReconciler, mockNow, SetupOptions{})
+		req, originalResources := setupNamespaceWithResources(ctx, sleepInfoName, namespace, sleepInfoReconciler, mockNow, setupOptions{})
 
 		assertContextInfo := AssertOperation{
 			testLogger:          testLogger,
@@ -355,7 +350,7 @@ var _ = Describe("SleepInfo Controller", func() {
 			req:                 req,
 			namespace:           namespace,
 			sleepInfoName:       sleepInfoName,
-			originalDeployments: originalDeployments,
+			originalDeployments: originalResources.deploymentList,
 		}
 		assertCorrectSleepOperation(assertContextInfo.withSchedule("2021-03-23T20:05:59.000Z").withRequeue(14*60 + 1))
 		assertCorrectSleepOperation(
@@ -368,8 +363,8 @@ var _ = Describe("SleepInfo Controller", func() {
 
 	It("reconcile - only sleep, wake up set to nil", func() {
 		namespace := "without-wake-up"
-		req, originalDeployments := setupNamespaceWithDeployments(ctx, sleepInfoName, namespace, sleepInfoReconciler, mockNow, SetupOptions{
-			UnsetWakeUpTime: true,
+		req, originalResources := setupNamespaceWithResources(ctx, sleepInfoName, namespace, sleepInfoReconciler, mockNow, setupOptions{
+			unsetWakeUpTime: true,
 		})
 
 		assertContextInfo := AssertOperation{
@@ -378,20 +373,18 @@ var _ = Describe("SleepInfo Controller", func() {
 			req:                 req,
 			namespace:           namespace,
 			sleepInfoName:       sleepInfoName,
-			originalDeployments: originalDeployments,
+			originalDeployments: originalResources.deploymentList,
 		}
 		assertCorrectSleepOperation(assertContextInfo.withSchedule("2021-03-23T20:05:59.000Z").withRequeue(59*60 + 1))
 		assertCorrectSleepOperation(assertContextInfo.withSchedule("2021-03-23T21:05:00.000Z").withRequeue(60 * 60))
 
 		By("re deploy", func() {
-			_, err := upsertDeployments(ctx, namespace, true)
-			Expect(err).NotTo(HaveOccurred())
+			upsertDeployments(ctx, namespace, true)
 
 			By("check replicas")
-			deployments, err := listDeployments(ctx, namespace)
-			Expect(err).NotTo(HaveOccurred())
+			deployments := listDeployments(ctx, namespace)
 			for _, deployment := range deployments {
-				originalDeployment := findDeployByName(originalDeployments, deployment.Name)
+				originalDeployment := findDeployByName(originalResources.deploymentList, deployment.Name)
 				Expect(deployment.Spec.Replicas).To(Equal(originalDeployment.Spec.Replicas))
 			}
 		})
@@ -417,9 +410,8 @@ var _ = Describe("SleepInfo Controller", func() {
 
 	It("reconcile - sleepinfo deployed when should be triggered", func() {
 		namespace := "immediately-triggered"
-		createSleepInfo(ctx, sleepInfoName, namespace, SetupOptions{})
-		originalDeployments, err := upsertDeployments(ctx, namespace, false)
-		Expect(err).NotTo(HaveOccurred())
+		createSleepInfo(ctx, sleepInfoName, namespace, setupOptions{})
+		originalDeployments := upsertDeployments(ctx, namespace, false)
 
 		req := reconcile.Request{
 			NamespacedName: types.NamespacedName{
@@ -438,8 +430,7 @@ var _ = Describe("SleepInfo Controller", func() {
 		Expect(err).NotTo(HaveOccurred())
 
 		By("replicas are set to 0 to all deployments", func() {
-			deployments, err := listDeployments(ctx, namespace)
-			Expect(err).NotTo(HaveOccurred())
+			deployments := listDeployments(ctx, namespace)
 			assertAllReplicasSetToZero(deployments, originalDeployments)
 		})
 
@@ -452,7 +443,7 @@ var _ = Describe("SleepInfo Controller", func() {
 
 	It("reconcile - create deployment between sleep and wake up", func() {
 		namespace := "create-deployment-between-sleep-and-wake-up"
-		req, originalDeployments := setupNamespaceWithDeployments(ctx, sleepInfoName, namespace, sleepInfoReconciler, mockNow, SetupOptions{})
+		req, originalResources := setupNamespaceWithResources(ctx, sleepInfoName, namespace, sleepInfoReconciler, mockNow, setupOptions{})
 
 		assertContextInfo := AssertOperation{
 			testLogger:          testLogger,
@@ -460,7 +451,7 @@ var _ = Describe("SleepInfo Controller", func() {
 			req:                 req,
 			namespace:           namespace,
 			sleepInfoName:       sleepInfoName,
-			originalDeployments: originalDeployments,
+			originalDeployments: originalResources.deploymentList,
 		}
 
 		assertCorrectSleepOperation(assertContextInfo.withSchedule("2021-03-23T20:05:59.000Z").withRequeue(14*60 + 1))
@@ -504,12 +495,11 @@ var _ = Describe("SleepInfo Controller", func() {
 			err := k8sClient.Create(ctx, deployToCreate.DeepCopy())
 			Expect(err).NotTo(HaveOccurred())
 
-			newOriginalDeployments := append(originalDeployments, deployToCreate)
+			newOriginalDeployments := append(originalResources.deploymentList, deployToCreate)
 			assertContextInfo.originalDeployments = newOriginalDeployments
 
 			By("check replicas")
-			deployments, err := listDeployments(ctx, namespace)
-			Expect(err).NotTo(HaveOccurred())
+			deployments := listDeployments(ctx, namespace)
 			for _, deployment := range deployments {
 				if deployment.Name == createdServiceName {
 					Expect(*deployment.Spec.Replicas).To(Equal(fiveReplicas))
@@ -524,8 +514,8 @@ var _ = Describe("SleepInfo Controller", func() {
 
 	It("reconcile - with deployments to exclude", func() {
 		namespace := "multiple-deployments-exclude"
-		req, originalDeployments := setupNamespaceWithDeployments(ctx, sleepInfoName, namespace, sleepInfoReconciler, mockNow, SetupOptions{
-			ExcludeRef: []kubegreenv1alpha1.ExcludeRef{
+		req, originalResources := setupNamespaceWithResources(ctx, sleepInfoName, namespace, sleepInfoReconciler, mockNow, setupOptions{
+			excludeRef: []kubegreenv1alpha1.ExcludeRef{
 				{
 					ApiVersion: "apps/v1",
 					Kind:       "Deployment",
@@ -545,7 +535,7 @@ var _ = Describe("SleepInfo Controller", func() {
 			req:                 req,
 			namespace:           namespace,
 			sleepInfoName:       sleepInfoName,
-			originalDeployments: originalDeployments,
+			originalDeployments: originalResources.deploymentList,
 			excludedDeployment:  []string{"service-1", "zero-replicas"},
 		}
 
@@ -554,198 +544,6 @@ var _ = Describe("SleepInfo Controller", func() {
 		assertCorrectSleepOperation(assertContextInfo.withSchedule("2021-03-23T21:05:00.000Z").withRequeue(15 * 60))
 	})
 })
-
-func createNamespace(ctx context.Context, name string) error {
-	namespace := &core.Namespace{
-		TypeMeta: metav1.TypeMeta{
-			Kind:       "Namespace",
-			APIVersion: "v1",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name: name,
-		},
-	}
-	return k8sClient.Create(ctx, namespace)
-}
-
-func upsertDeployments(ctx context.Context, namespace string, updateIfAlreadyCreated bool) ([]appsv1.Deployment, error) {
-	var threeReplicas int32 = 3
-	var oneReplica int32 = 1
-	var zeroReplicas int32 = 0
-	deployments := []appsv1.Deployment{
-		{
-			TypeMeta: metav1.TypeMeta{
-				Kind:       "Deployment",
-				APIVersion: "apps/v1",
-			},
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      "service-1",
-				Namespace: namespace,
-			},
-			Spec: appsv1.DeploymentSpec{
-				Replicas: &threeReplicas,
-				Selector: &metav1.LabelSelector{
-					MatchLabels: map[string]string{
-						"app": "service-1",
-					},
-				},
-				Template: core.PodTemplateSpec{
-					ObjectMeta: metav1.ObjectMeta{
-						Labels: map[string]string{
-							"app": "service-1",
-						},
-					},
-					Spec: core.PodSpec{
-						Containers: []core.Container{
-							{
-								Name:  "c1",
-								Image: "davidebianchi/echo-service",
-							},
-						},
-					},
-				},
-			},
-		},
-		{
-			TypeMeta: metav1.TypeMeta{
-				Kind:       "Deployment",
-				APIVersion: "apps/v1",
-			},
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      "service-2",
-				Namespace: namespace,
-			},
-			Spec: appsv1.DeploymentSpec{
-				Replicas: &oneReplica,
-				Selector: &metav1.LabelSelector{
-					MatchLabels: map[string]string{
-						"app": "service-2",
-					},
-				},
-				Template: core.PodTemplateSpec{
-					ObjectMeta: metav1.ObjectMeta{
-						Labels: map[string]string{
-							"app": "service-2",
-						},
-					},
-					Spec: core.PodSpec{
-						Containers: []core.Container{
-							{
-								Name:  "c2",
-								Image: "davidebianchi/echo-service",
-							},
-						},
-					},
-				},
-			},
-		},
-		{
-			TypeMeta: metav1.TypeMeta{
-				Kind:       "Deployment",
-				APIVersion: "apps/v1",
-			},
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      "zero-replicas",
-				Namespace: namespace,
-			},
-			Spec: appsv1.DeploymentSpec{
-				Replicas: &zeroReplicas,
-				Selector: &metav1.LabelSelector{
-					MatchLabels: map[string]string{
-						"app": "zero-replicas",
-					},
-				},
-				Template: core.PodTemplateSpec{
-					ObjectMeta: metav1.ObjectMeta{
-						Labels: map[string]string{
-							"app": "zero-replicas",
-						},
-					},
-					Spec: core.PodSpec{
-						Containers: []core.Container{
-							{
-								Name:  "zero",
-								Image: "davidebianchi/echo-service",
-							},
-						},
-					},
-				},
-			},
-		},
-		{
-			TypeMeta: metav1.TypeMeta{
-				Kind:       "Deployment",
-				APIVersion: "apps/v1",
-			},
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      "zero-replicas-annotation",
-				Namespace: namespace,
-				Annotations: map[string]string{
-					lastScheduleKey: "2021-03-23T00:00:00.000Z",
-				},
-			},
-			Spec: appsv1.DeploymentSpec{
-				Replicas: &zeroReplicas,
-				Selector: &metav1.LabelSelector{
-					MatchLabels: map[string]string{
-						"app": "zero-replicas-annotation",
-					},
-				},
-				Template: core.PodTemplateSpec{
-					ObjectMeta: metav1.ObjectMeta{
-						Labels: map[string]string{
-							"app": "zero-replicas-annotation",
-						},
-					},
-					Spec: core.PodSpec{
-						Containers: []core.Container{
-							{
-								Name:  "zero",
-								Image: "davidebianchi/echo-service",
-							},
-						},
-					},
-				},
-			},
-		},
-	}
-	for _, deployment := range deployments {
-		var deploymentAlreadyExists bool
-		d := appsv1.Deployment{}
-		if updateIfAlreadyCreated {
-			err := k8sClient.Get(ctx, types.NamespacedName{
-				Name:      deployment.Name,
-				Namespace: namespace,
-			}, &d)
-			if err == nil {
-				deploymentAlreadyExists = true
-			}
-		}
-		if deploymentAlreadyExists {
-			patch := client.MergeFrom(d.DeepCopy())
-			d.Spec.Replicas = deployment.Spec.Replicas
-			if err := k8sClient.Patch(ctx, &d, patch); err != nil {
-				return nil, fmt.Errorf("error updating deployment: %s %s", err, deployment.Name)
-			}
-		} else {
-			if err := k8sClient.Create(ctx, &deployment); err != nil {
-				return nil, fmt.Errorf("error %s creating deployment 1", err)
-			}
-		}
-	}
-	return deployments, nil
-}
-
-func listDeployments(ctx context.Context, namespace string) ([]appsv1.Deployment, error) {
-	deployments := appsv1.DeploymentList{}
-	err := k8sClient.List(ctx, &deployments, &client.ListOptions{
-		Namespace: namespace,
-	})
-	if err != nil {
-		return nil, err
-	}
-	return deployments.Items, nil
-}
 
 type mockClock struct {
 	now string
@@ -818,8 +616,7 @@ func assertCorrectSleepOperation(assert AssertOperation) {
 	Expect(err).NotTo(HaveOccurred())
 
 	By("replicas are set to 0 to all deployments set to sleep", func() {
-		deployments, err := listDeployments(assert.ctx, assert.namespace)
-		Expect(err).NotTo(HaveOccurred())
+		deployments := listDeployments(assert.ctx, assert.namespace)
 		if len(assert.excludedDeployment) == 0 {
 			assertAllReplicasSetToZero(deployments, assert.originalDeployments)
 		} else {
@@ -893,8 +690,7 @@ func assertCorrectWakeUpOperation(assert AssertOperation) {
 	Expect(err).NotTo(HaveOccurred())
 
 	By("deployment replicas correctly waked up", func() {
-		deployments, err := listDeployments(assert.ctx, assert.namespace)
-		Expect(err).NotTo(HaveOccurred())
+		deployments := listDeployments(assert.ctx, assert.namespace)
 		for _, deployment := range deployments {
 			originalDeployment := findDeployByName(assert.originalDeployments, deployment.Name)
 			Expect(deployment.Spec.Replicas).To(Equal(originalDeployment.Spec.Replicas))
@@ -925,122 +721,4 @@ func assertCorrectWakeUpOperation(assert AssertOperation) {
 			RequeueAfter: assert.expectedNextRequeue,
 		}))
 	})
-}
-
-type SetupOptions struct {
-	ExcludeRef      []kubegreenv1alpha1.ExcludeRef
-	UnsetWakeUpTime bool
-}
-
-func setupNamespaceWithDeployments(ctx context.Context, sleepInfoName, namespace string, reconciler SleepInfoReconciler, now string, opts SetupOptions) (ctrl.Request, []appsv1.Deployment) {
-	createSleepInfo(ctx, sleepInfoName, namespace, opts)
-
-	By("create deployments")
-	originalDeployments, err := upsertDeployments(ctx, namespace, false)
-	Expect(err).NotTo(HaveOccurred())
-
-	req := reconcile.Request{
-		NamespacedName: types.NamespacedName{
-			Name:      sleepInfoName,
-			Namespace: namespace,
-		},
-	}
-	result, err := reconciler.Reconcile(ctx, req)
-	Expect(err).NotTo(HaveOccurred())
-
-	By("is requeued after correct duration", func() {
-		Expect(result).Should(Equal(ctrl.Result{
-			RequeueAfter: sleepRequeue(now),
-		}))
-	})
-
-	By("replicas not changed", func() {
-		deploymentsNotChanged, err := listDeployments(ctx, namespace)
-		Expect(err).NotTo(HaveOccurred())
-		for i, deployment := range deploymentsNotChanged {
-			Expect(*deployment.Spec.Replicas).To(Equal(*originalDeployments[i].Spec.Replicas))
-		}
-	})
-
-	return req, originalDeployments
-}
-
-func createSleepInfo(ctx context.Context, sleepInfoName, namespace string, opts SetupOptions) kubegreenv1alpha1.SleepInfo {
-	var (
-		timeout  = time.Second * 10
-		interval = time.Millisecond * 250
-	)
-
-	Expect(createNamespace(ctx, namespace)).NotTo(HaveOccurred())
-	sleepInfo := &kubegreenv1alpha1.SleepInfo{
-		TypeMeta: metav1.TypeMeta{
-			Kind:       "SleepInfo",
-			APIVersion: "kube-green.com/v1alpha1",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      sleepInfoName,
-			Namespace: namespace,
-		},
-		Spec: kubegreenv1alpha1.SleepInfoSpec{
-			Weekdays:   "*",
-			SleepTime:  "*:05", // every 5 minute
-			WakeUpTime: "*:20", // every 20 minute
-		},
-	}
-	if opts.UnsetWakeUpTime {
-		sleepInfo.Spec.WakeUpTime = ""
-	}
-	if len(opts.ExcludeRef) != 0 {
-		sleepInfo.Spec.ExcludeRef = opts.ExcludeRef
-	}
-
-	Expect(k8sClient.Create(ctx, sleepInfo)).Should(Succeed())
-
-	sleepInfoLookupKey := types.NamespacedName{Name: sleepInfoName, Namespace: namespace}
-	createdSleepInfo := &kubegreenv1alpha1.SleepInfo{}
-
-	// We'll need to retry getting this newly created SleepInfo, given that creation may not immediately happen.
-	Eventually(func() bool {
-		err := k8sClient.Get(ctx, sleepInfoLookupKey, createdSleepInfo)
-		Expect(err).NotTo(HaveOccurred())
-		return true
-	}, timeout, interval).Should(BeTrue())
-
-	return *createdSleepInfo
-}
-
-func sleepRequeue(now string) time.Duration {
-	parsedTime, _ := time.Parse(time.RFC3339, now)
-	hour := parsedTime.Hour()
-	if parsedTime.Minute() > 5 {
-		hour += 1
-	}
-	return time.Duration((time.Date(2021, time.March, 23, hour, 5, 0, 0, time.UTC).UnixNano() - parsedTime.UnixNano()))
-}
-
-func wakeUpRequeue(now string) time.Duration {
-	parsedTime, _ := time.Parse(time.RFC3339, now)
-	hour := parsedTime.Hour()
-	if parsedTime.Minute() > 20 {
-		hour += 1
-	}
-	return time.Duration((time.Date(2021, time.March, 23, hour, 20, 0, 0, time.UTC).UnixNano() - parsedTime.UnixNano()))
-}
-
-func findDeployByName(deployments []appsv1.Deployment, nameToFind string) *appsv1.Deployment {
-	for _, deployment := range deployments {
-		if deployment.Name == nameToFind {
-			return deployment.DeepCopy()
-		}
-	}
-	return nil
-}
-
-func contains(s []string, v string) bool {
-	for _, a := range s {
-		if a == v {
-			return true
-		}
-	}
-	return false
 }
